requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
asyncpg==0.30.0
bcrypt==4.2.1
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
ecdsa==0.19.0
fastapi==0.115.8
GeoAlchemy2==0.17.0
googlemaps==4.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isort==6.0.0
jiter==0.8.2
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==2.2.3
openai==1.62.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.4
python-dotenv==1.0.1
python-jose==3.4.0
requests==2.32.3
rich==13.9.4
rsa==4.9
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
